requirements.txt
django>=1.8.3,<1.9
django-extensions>=1.5.2
https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a3.tar.gz
django-bootstrap3>=5.3.1,<6.0
django-ldap-email-auth==0.1.2
webtemplate-dpaw==0.3
openpyxl>=2.2.1
Werkzeug>=0.10.4
django-reversion>=1.8.7
Fabric>=1.10.1
jinja2>=2.7.3
django-tastypie==0.12.2
django-grappelli>=2.6.5
django-envelope>=1.0
django-braces>=1.8.0
django-debug-toolbar>=1.3.0
# Testing requirements
coverage==3.7.1
flake8==2.4.0
mixer==5.1.2
